refactor(pipeline): decompose verify_cli.py to ≤80 lines (#564)
verify_cli.py was 113 lines, over the 80-line pipeline cap. Split into
three responsibility modules under stages/common/:

- verify_context.py (26) — VerifyContext frozen dataclass
- verify_args.py (64) — parse_verify_args(); reuses _derive_zone() and
  _read_erp_user() from config_helpers instead of the prior inline
  duplicates (the #563-flagged duplication)
- verify_cli.py (35) — thin facade: print_results() + re-exports
  parse_verify_args + VerifyContext

The facade keeps all 10 stage verify.py import sites unchanged. Pure
refactor, no behavioral change: parse_verify_args derivation verified
identical for dev02; live e2e stage_5 (3/3) and stage_7 (5/5) pass.

fixes #564
