+++ tools/size_baselines.json
@@ -128,7 +128,9 @@
   "tools/pipeline/stages/common/ssh.py": 83,
   "tools/pipeline/stages/common/types.py": 57,
   "tools/pipeline/stages/common/verify_ansible_output.py": 46,
-  "tools/pipeline/stages/common/verify_cli.py": 113,
+  "tools/pipeline/stages/common/verify_args.py": 64,
+  "tools/pipeline/stages/common/verify_cli.py": 35,
+  "tools/pipeline/stages/common/verify_context.py": 26,
   "tools/pipeline/stages/destroy.py": 1,
   "tools/pipeline/stages/env_cloudstack.py": 1,
   "tools/pipeline/stages/env_kvm.py": 40,
